Fix handling of redirect location
WP7 fails to parse absolute URLs starting with a "/" and it turns them into "file:///" URIs. So, now we try first to parse relative URLs based on NetMetrix's host, and if that fails, we go for absolute URL parsing. Absolute URLs need to be supported because NetMetrix test environment uses cross-domain redirects (but not the live environment).